+++ tests/check_serialization.cpp
@@ -19,13 +19,13 @@ TEST_CASE("Testing serialized size computation")
     CHECK(Serial::getSize((double)18.0) == sizeof(double));
 
     vector<int> vectorOfInts{1, 2, 3, 4};
-    CHECK(Serial::getSize(vectorOfInts) == vectorOfInts.size() * sizeof(int) + sizeof(size_t));
+    CHECK(Serial::getSize(vectorOfInts) == vectorOfInts.size() * sizeof(int) + sizeof(uint32_t));
 
     deque<float> dequeOfFloat{3.14159f, 42.f, 2.71828f};
-    CHECK(Serial::getSize(dequeOfFloat) == dequeOfFloat.size() * sizeof(float) + sizeof(size_t));
+    CHECK(Serial::getSize(dequeOfFloat) == dequeOfFloat.size() * sizeof(float) + sizeof(uint32_t));
 
     std::string someText{"So long, and thanks for the fish!"};
-    CHECK(Serial::getSize(someText) == someText.size() * sizeof(char) + sizeof(size_t));
+    CHECK(Serial::getSize(someText) == someText.size() * sizeof(char) + sizeof(uint32_t));
 
     CHECK(Serial::getSize(chrono::system_clock::time_point(chrono::duration<int64_t, std::milli>(123456))) == sizeof(int64_t));
 }
@@ -38,7 +38,7 @@ TEST_CASE("Testing serialized size computation for containers of containers")
         for (uint32_t i = 0; i < 4; ++i)
             data.push_back(vector<int>({2, 4, 8, 16}));
 
-        CHECK(Serial::getSize(data) == sizeof(size_t) + data.size() * (sizeof(size_t) + data[0].size() * sizeof(int)));
+        CHECK(Serial::getSize(data) == sizeof(uint32_t) + data.size() * (sizeof(uint32_t) + data[0].size() * sizeof(int)));
     }
 
     {
@@ -46,7 +46,7 @@ TEST_CASE("Testing serialized size computation for containers of containers")
         for (uint32_t i = 0; i < 4; ++i)
             data.push_back(deque<int>({2, 4, 8, 16}));
 
-        CHECK(Serial::getSize(data) == sizeof(size_t) + data.size() * (sizeof(size_t) + data[0].size() * sizeof(int)));
+        CHECK(Serial::getSize(data) == sizeof(uint32_t) + data.size() * (sizeof(uint32_t) + data[0].size() * sizeof(int)));
     }
 }
 
@@ -55,7 +55,7 @@ TEST_CASE("Testing serialized size computation for tuples")
 {
     auto testString = string("Show me the money");
     auto data = make_tuple(3.1415f, 42, testString);
-    CHECK(Serial::getSize(data) == sizeof(float) + sizeof(int) + sizeof(size_t) + testString.size());
+    CHECK(Serial::getSize(data) == sizeof(float) + sizeof(int) + sizeof(uint32_t) + testString.size());
 }
 
 /*************/
@@ -76,9 +76,9 @@ TEST_CASE("Testing elementary serialization")
     bufferPtr += sizeof(float);
     CHECK(*reinterpret_cast<double*>(bufferPtr) == 2.71828);
     bufferPtr += sizeof(double);
-    auto stringLength = *reinterpret_cast<size_t*>(bufferPtr);
+    auto stringLength = *reinterpret_cast<uint32_t*>(bufferPtr);
     CHECK(stringLength == testString.size());
-    bufferPtr += sizeof(size_t);
+    bufferPtr += sizeof(uint32_t);
     CHECK(string(reinterpret_cast<char*>(bufferPtr), stringLength) == testString);
     bufferPtr += sizeof(char) * testString.size();
     CHECK(*reinterpret_cast<int64_t*>(bufferPtr) == 123456);
@@ -92,7 +92,7 @@ TEST_CASE("Testing serialization of iterable containers")
         vector<int> data{1, 1, 2, 3, 5, 8};
         Serial::serialize(data, buffer);
 
-        auto bufferPtr = reinterpret_cast<int*>(buffer.data() + sizeof(size_t));
+        auto bufferPtr = reinterpret_cast<int*>(buffer.data() + sizeof(uint32_t));
         CHECK(buffer.size() == Serial::getSize(data));
         for (uint32_t i = 0; i < data.size(); ++i)
             CHECK(data[i] == bufferPtr[i]);
@@ -102,7 +102,7 @@ TEST_CASE("Testing serialization of iterable containers")
         vector<uint8_t> buffer;
         deque<float> data{3.14159f, 2.71828f};
         Serial::serialize(data, buffer);
-        auto bufferPtr = reinterpret_cast<float*>(buffer.data() + sizeof(size_t));
+        auto bufferPtr = reinterpret_cast<float*>(buffer.data() + sizeof(uint32_t));
         CHECK(buffer.size() == Serial::getSize(data));
         for (uint32_t i = 0; i < data.size(); ++i)
             CHECK(data[i] == bufferPtr[i]);
@@ -122,9 +122,9 @@ TEST_CASE("Testing serialization of tuples")
     bufferPtr += sizeof(float);
     CHECK(*reinterpret_cast<int*>(bufferPtr) == 42);
     bufferPtr += sizeof(int);
-    auto stringLength = *reinterpret_cast<size_t*>(bufferPtr);
+    auto stringLength = *reinterpret_cast<uint32_t*>(bufferPtr);
     CHECK(stringLength == testString.size());
-    bufferPtr += sizeof(size_t);
+    bufferPtr += sizeof(uint32_t);
     CHECK(string(reinterpret_cast<char*>(bufferPtr), stringLength) == testString);
 }
